+++ centreon/www/include/options/accessLists/groupsACL/DB-Func.php
@@ -165,15 +165,14 @@ function deleteGroupInDB($groups = [])
         $dbResult->bindValue('aclGroupId', $key, PDO::PARAM_INT);
         $dbResult->execute();
         $row = $dbResult->fetch();
+        if ($row === false) {
+            continue;
+        }
+
         $dbResult = $pearDB->prepare('DELETE FROM acl_groups WHERE acl_group_id = :aclGroupId');
         $dbResult->bindValue('aclGroupId', $key, PDO::PARAM_INT);
         $dbResult->execute();
-        $centreon->CentreonLogAction->insertLog(
-            'access group',
-            (int) $key,
-            $row !== false ? $row['acl_group_name'] : "id:{$key}",
-            'd'
-        );
+        $centreon->CentreonLogAction->insertLog('access group', (int) $key, $row['acl_group_name'], 'd');
     }
 }
